add build GitHub workflow
This is a matrix strategy build which is probably the most useful default workflow for end-users.

Should we include default board/shield values out-of-the-box (currently nice_nano + corne_left/right) or add gaps or template-placeholders to encourage users to implement this part of the build themselves?  If application cloning/creation is facilitated as part of the ZMK website, then we could auto add/commit the board/shield for new users.

We could perhaps include several types of workflows which module authors can activate as needed, such as linting or formatting.

We could also provide an additional build workflow that only runs in the base repository protected by a ${{ github.repository }} guard on the URL or name.  This would let CI check the base version.